optimize hero img to 1800px
